+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ## Changelog
-- fix `mafglib.forge.mixins.json`
+- rewrite `ForgeUtils`